Working build for file exchange

+++ daemon/folder/transfer/downloader/ChunkFileBuilder.cpp
@@ -36,46 +36,45 @@ Q_DECLARE_LOGGING_CATEGORY(log_downloader)
 
 /* ChunkFileBuilderFdPool */
 QFile* ChunkFileBuilderFdPool::getFile(QString path, bool release) {
-	QFile* f = !release ? opened_files_[path] : opened_files_.take(path);
-	if(f)
-		return f;
+  QFile* f = !release ? opened_files_[path] : opened_files_.take(path);
+  if (f) return f;
 
-	f = new QFile(path);
-	if(! f->open(QIODevice::ReadWrite))
-		qCWarning(log_downloader) << "Could not open" << path << "Error:" << f->errorString();
-	opened_files_.insert(path, f);
+  f = new QFile(path);
+  if (!f->open(QIODevice::ReadWrite))
+    qCWarning(log_downloader) << "Could not open" << path << "Error:" << f->errorString();
+  opened_files_.insert(path, f);
 
-	return f;
+  return f;
 }
 
 /* ChunkFileBuilder */
-ChunkFileBuilder::ChunkFileBuilder(const QString& system_path, const QByteArray& ct_hash, quint32 size) : file_map_(size) {
-	chunk_location_ = system_path + "/incomplete-" + QString::fromLatin1(toBase32(ct_hash));
+ChunkFileBuilder::ChunkFileBuilder(
+    const QString& system_path, const QByteArray& ct_hash, quint32 size)
+    : file_map_(size) {
+  chunk_location_ = system_path + "/incomplete-" + QString::fromLatin1(toBase32(ct_hash));
 
-	QFile f(chunk_location_);
-	f.open(QIODevice::WriteOnly | QIODevice::Truncate);
-	f.resize(size);
+  QFile f(chunk_location_);
+  f.open(QIODevice::WriteOnly | QIODevice::Truncate);
+  f.resize(size);
 }
 
 ChunkFileBuilder::~ChunkFileBuilder() {
-	if(! chunk_location_.isEmpty())
-		QFile::remove(chunk_location_);
+  if (!chunk_location_.isEmpty()) QFile::remove(chunk_location_);
 }
 
-QFile* ChunkFileBuilder::release_chunk() {
-	QFile* f = ChunkFileBuilderFdPool::get_instance()->getFile(chunk_location_, true);
-	chunk_location_.clear();
-	return f;
+QFile* ChunkFileBuilder::releaseChunk() {
+  QFile* f = ChunkFileBuilderFdPool::get_instance()->getFile(chunk_location_, true);
+  chunk_location_.clear();
+  return f;
 }
 
-void ChunkFileBuilder::put_block(quint32 offset, const QByteArray& content) {
-	auto inserted = file_map_.insert({offset, content.size()}).second;
-	if(inserted) {
-		QFile* f = ChunkFileBuilderFdPool::get_instance()->getFile(chunk_location_);
-		if(f->pos() != offset)
-			f->seek(offset);
-		f->write(content);
-	}
+void ChunkFileBuilder::putBlock(quint32 offset, const QByteArray& content) {
+  auto inserted = file_map_.insert({offset, content.size()}).second;
+  if (inserted) {
+    QFile* f = ChunkFileBuilderFdPool::get_instance()->getFile(chunk_location_);
+    if (f->pos() != offset) f->seek(offset);
+    f->write(content);
+  }
 }
 
 } /* namespace librevault */
